Split into TCP and UDP Packet server, add abstract class
